+++ examples/readme.rs
@@ -1,17 +1,22 @@
 fn main() {
-    use ansi_term::{ANSIStrings, Color, Style};
+    use anstyle::{AnsiColor, Style};
     use tui_markup::{compile, compile_with, generator::ANSIStringsGenerator};
 
     // Parse markup into some final result for showing
     let result = compile::<ANSIStringsGenerator>("You got a <yellow Coin>").unwrap();
     // Show it
-    println!("{}", ANSIStrings(&result));
+    println!("{}", result);
 
     // With custom tag
     let generator = ANSIStringsGenerator::new(|tag: &str| match tag {
-        "keyboard" => Some(Style::default().fg(Color::Blue).on(Color::Black).bold()),
+        "keyboard" => Some(
+            Style::new()
+                .fg_color(Some(AnsiColor::Blue.into()))
+                .bg_color(Some(AnsiColor::Black.into()))
+                .bold(),
+        ),
         _ => None,
     });
     let result = compile_with("Press <keyboard Space> to jump", generator).unwrap();
-    println!("{}", ANSIStrings(&result));
+    println!("{}", result);
 }