+++ src/generator/ratatui/tag.rs
@@ -1,4 +1,4 @@
-use ratatui::style::{Color, Modifier, Style};
+use ratatui_core::style::{Color, Modifier, Style};
 
 use crate::{
     generator::{